Added a period.
I can't take it. I have to. Somebody stop me. Before it's too late. Before I make a commit with no changes except for the addition of a single period. No. Not like this. I can feel myself being drawn to the 'commit changes' button. So green; so clickable. I have to stop myself. I can't do this. But the period! The gap where a period should have been taunts me. Missing a single period? How could I do that? But this commit, is it worth it? I have to do it. I can't leave a period missing just like that. I'm doing it. I wont deprive a sentence of its most important part; its final farewell to the reader; its definitive conclusion